# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bff_elearning",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "PAYLOAD_TOO_LARGE",
                  "UNPROCESSABLE_ENTITY",
                  "TOO_MANY_REQUESTS",
                  "INTERNAL_ERROR",
                  "BAD_GATEWAY",
                  "SERVICE_UNAVAILABLE",
                  "GATEWAY_TIMEOUT"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/elearning/courses/{courseId}/contents/{contentId}/complete": {
      "post": {
        "tags": [
          "E-learning"
        ],
        "summary": "Marks a content as completed",
        "description": "Records the completion in the E-learning API, which tracks progress per chapter: completing a content completes its whole chapter. Returns the updated progress, chapters, chapter and content. `completed: false` answers 501 (the E-learning API cannot undo a completion).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9]\\d{0,14}$",
              "description": "Course (E-learning API formation) identifier",
              "example": "4"
            },
            "required": true,
            "description": "Course (E-learning API formation) identifier",
            "name": "courseId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[1-9]\\d{0,14}$",
              "description": "Content (E-learning API attachment) identifier",
              "example": "27"
            },
            "required": true,
            "description": "Content (E-learning API attachment) identifier",
            "name": "contentId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "chapterId": {
                    "type": "string",
                    "pattern": "^[1-9]\\d{0,14}$",
                    "description": "Chapter (E-learning API module) identifier",
                    "example": "11"
                  },
                  "completed": {
                    "type": "boolean",
                    "description": "Only `true` is supported: the E-learning API cannot mark a module as not completed (501).",
                    "example": true
                  }
                },
                "required": [
                  "chapterId",
                  "completed"
                ],
                "description": "Content completion. Progress is tracked per chapter by the E-learning API: completing a content completes its whole chapter."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Progress updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "progress": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 100
                    },
                    "completedRequiredContents": {
                      "type": "number",
                      "minimum": 0
                    },
                    "totalRequiredContents": {
                      "type": "number",
                      "minimum": 0
                    },
                    "completedChapters": {
                      "type": "number",
                      "minimum": 0
                    },
                    "totalChapters": {
                      "type": "number",
                      "minimum": 0
                    },
                    "completed": {
                      "type": "boolean"
                    },
                    "chapters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "accueil-1"
                          },
                          "title": {
                            "type": "string",
                            "example": "Municipal organisation"
                          },
                          "description": {
                            "type": "string"
                          },
                          "duration": {
                            "type": "string",
                            "example": "25 min"
                          },
                          "completed": {
                            "type": "boolean"
                          },
                          "active": {
                            "type": "boolean"
                          },
                          "contents": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "example": "accueil-1-video"
                                },
                                "title": {
                                  "type": "string",
                                  "example": "Présentation des directions et services"
                                },
                                "type": {
                                  "type": "string",
                                  "enum": [
                                    "video",
                                    "pdf",
                                    "document",
                                    "link",
                                    "quiz",
                                    "audio",
                                    "other"
                                  ],
                                  "description": "Type de contenu pédagogique",
                                  "example": "video"
                                },
                                "description": {
                                  "type": "string"
                                },
                                "duration": {
                                  "type": "string",
                                  "example": "12 min"
                                },
                                "fileName": {
                                  "type": "string",
                                  "example": "organigramme-mairie360.pdf"
                                },
                                "href": {
                                  "type": "string",
                                  "example": "/documents/organigramme-mairie360.pdf"
                                },
                                "completed": {
                                  "type": "boolean"
                                },
                                "required": {
                                  "type": "boolean"
                                }
                              },
                              "required": [
                                "id",
                                "title",
                                "type"
                              ],
                              "description": "Contenu pédagogique d’un chapitre"
                            }
                          }
                        },
                        "required": [
                          "id",
                          "title"
                        ],
                        "description": "Chapitre d’une formation"
                      }
                    },
                    "chapter": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "accueil-1"
                        },
                        "title": {
                          "type": "string",
                          "example": "Municipal organisation"
                        },
                        "description": {
                          "type": "string"
                        },
                        "duration": {
                          "type": "string",
                          "example": "25 min"
                        },
                        "completed": {
                          "type": "boolean"
                        },
                        "active": {
                          "type": "boolean"
                        },
                        "contents": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "example": "accueil-1-video"
                              },
                              "title": {
                                "type": "string",
                                "example": "Présentation des directions et services"
                              },
                              "type": {
                                "type": "string",
                                "enum": [
                                  "video",
                                  "pdf",
                                  "document",
                                  "link",
                                  "quiz",
                                  "audio",
                                  "other"
                                ],
                                "description": "Type de contenu pédagogique",
                                "example": "video"
                              },
                              "description": {
                                "type": "string"
                              },
                              "duration": {
                                "type": "string",
                                "example": "12 min"
                              },
                              "fileName": {
                                "type": "string",
                                "example": "organigramme-mairie360.pdf"
                              },
                              "href": {
                                "type": "string",
                                "example": "/documents/organigramme-mairie360.pdf"
                              },
                              "completed": {
                                "type": "boolean"
                              },
                              "required": {
                                "type": "boolean"
                              }
                            },
                            "required": [
                              "id",
                              "title",
                              "type"
                            ],
                            "description": "Contenu pédagogique d’un chapitre"
                          }
                        }
                      },
                      "required": [
                        "id",
                        "title"
                      ],
                      "description": "Chapitre d’une formation"
                    },
                    "content": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "accueil-1-video"
                        },
                        "title": {
                          "type": "string",
                          "example": "Présentation des directions et services"
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "video",
                            "pdf",
                            "document",
                            "link",
                            "quiz",
                            "audio",
                            "other"
                          ],
                          "description": "Type de contenu pédagogique",
                          "example": "video"
                        },
                        "description": {
                          "type": "string"
                        },
                        "duration": {
                          "type": "string",
                          "example": "12 min"
                        },
                        "fileName": {
                          "type": "string",
                          "example": "organigramme-mairie360.pdf"
                        },
                        "href": {
                          "type": "string",
                          "example": "/documents/organigramme-mairie360.pdf"
                        },
                        "completed": {
                          "type": "boolean"
                        },
                        "required": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "title",
                        "type"
                      ],
                      "description": "Contenu pédagogique d’un chapitre"
                    }
                  },
                  "required": [
                    "progress",
                    "completedRequiredContents",
                    "totalRequiredContents",
                    "completedChapters",
                    "totalChapters",
                    "completed",
                    "chapters",
                    "chapter",
                    "content"
                  ],
                  "description": "Progression mise à jour après complétion d’un contenu"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Course, chapter or content not found, or the caller is not enrolled in the course",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not available: no upstream service stores this data yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "BFF User or an upstream API is unreachable, not configured, failed or answered an unusable body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}